[DIP/examples] Move images and kernels.h outside ConvOpt
